% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{DEBUG}
\alias{DEBUG}
\title{Debug log level constant}
\format{
An object of class \code{integer} of length 1.
}
\usage{
DEBUG
}
\description{
Numeric constant representing the DEBUG log level with value 8.
Used for logging detailed information for debugging purposes.
}
\keyword{datasets}
