% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{.logger_env}
\alias{.logger_env}
\title{Logger environment for storing configuration}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.logger_env
}
\description{
Internal environment to store logger configuration variables
like the current threshold level. This prevents issues with
locked bindings in package namespaces.
}
\keyword{datasets}
