% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{ERROR}
\alias{ERROR}
\title{Error log level constant}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ERROR
}
\description{
Numeric constant representing the ERROR log level with value 2.
Used for logging error conditions that don't necessarily terminate the application.
}
\keyword{datasets}
