% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{FATAL}
\alias{FATAL}
\title{Fatal log level constant}
\format{
An object of class \code{integer} of length 1.
}
\usage{
FATAL
}
\description{
Numeric constant representing the FATAL log level with value 1.
Used for logging critical errors that may cause the application to terminate.
}
\keyword{datasets}
