% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_files.R
\name{getAppFilePath}
\alias{getAppFilePath}
\title{Get App File Path (Deprecated)}
\usage{
getAppFilePath(appSpecUserFileSettingId, fallbackToProvidedFiles = TRUE)
}
\arguments{
\item{appSpecUserFileSettingId}{Character string. The ID of the requested set of app-files}

\item{fallbackToProvidedFiles}{Logical. Should the function fallback to bundled directory? Default is \code{TRUE}.}
}
\value{
Character string containing the directory path, or \code{NULL} if not found.
}
\description{
\strong{DEPRECATED!} This function handles app-settings of type \code{LOCAL_FILE}.
This setting type was deprecated. Please migrate from \code{LOCAL_FILE} to
\code{USER_FILE} app-setting type and use \code{\link{getAuxiliaryFilePath}} instead.
}
\keyword{internal}
