% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_app_files.R
\name{getUploadDirOrFallbackDir}
\alias{getUploadDirOrFallbackDir}
\title{Get Upload Directory or Fallback Directory}
\usage{
getUploadDirOrFallbackDir(
  appSpecUserFileSettingId,
  fallbackToProvidedFiles,
  userUploadDir,
  appDevFallbackDir
)
}
\arguments{
\item{appSpecUserFileSettingId}{Character string. The ID of the requested set of app-files}

\item{fallbackToProvidedFiles}{Logical. Should the function fallback to bundled directory?}

\item{userUploadDir}{Character string. Path to the user upload directory}

\item{appDevFallbackDir}{Character string. Path to the app developer fallback directory}
}
\value{
Character string containing the directory path, or \code{NULL} if no valid directory is found.
}
\description{
Internal function that provides the path to the directory of an auxiliary file.
This function should not be used by app developers directly.
Use \code{\link{getAuxiliaryFilePath}} to get the path to the file instead of just the parent directory.
}
\details{
The function checks for user-uploaded files first. If none are found and
\code{fallbackToProvidedFiles} is \code{TRUE}, it falls back to app developer
provided files. If no files are found in either location, it returns \code{NULL}.
}
\keyword{internal}
