% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{INFO}
\alias{INFO}
\title{Information log level constant}
\format{
An object of class \code{integer} of length 1.
}
\usage{
INFO
}
\description{
Numeric constant representing the INFO log level with value 6.
Used for logging general informational messages about application flow.
}
\keyword{datasets}
