% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.debug}
\alias{logger.debug}
\title{Log a debug level message}
\usage{
logger.debug(msg, ...)
}
\arguments{
\item{msg}{A character string containing the log message, supports sprintf formatting}

\item{...}{Additional arguments for sprintf-style message formatting}
}
\description{
Logs a message at the DEBUG level, used for detailed information that is
typically only of interest when diagnosing problems.
}
\examples{
logger.debug("Processing item \%d of \%d", 5, 10)
logger.debug("Current state: \%s", current_state)

}
\seealso{
\code{\link{logger.log_level}}, \code{\link{DEBUG}}
}
