% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.error}
\alias{logger.error}
\title{Log an error message}
\usage{
logger.error(msg, ...)
}
\arguments{
\item{msg}{A character string containing the log message, supports sprintf formatting}

\item{...}{Additional arguments for sprintf-style message formatting}
}
\description{
Logs a message at the ERROR level, used for error events that might still
allow the application to continue running.
}
\examples{
logger.error("Failed to connect to database: \%s", "foo")
logger.error("Invalid input provided: \%s", "invalid")

}
\seealso{
\code{\link{logger.log_level}}, \code{\link{ERROR}}
}
