% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.fatal}
\alias{logger.fatal}
\title{Log a fatal error message}
\usage{
logger.fatal(msg, ...)
}
\arguments{
\item{msg}{A character string containing the log message, supports sprintf formatting}

\item{...}{Additional arguments for sprintf-style message formatting}
}
\description{
Logs a message at the FATAL level, used for very severe error events that
might lead the application to abort or terminate.
}
\examples{
logger.fatal("Critical system error: \%s", system_error)
logger.fatal("Application cannot continue due to: \%s", fatal_condition)

}
\seealso{
\code{\link{logger.log_level}}, \code{\link{FATAL}}
}
