% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.info}
\alias{logger.info}
\title{Log an informational message}
\usage{
logger.info(msg, ...)
}
\arguments{
\item{msg}{A character string containing the log message, supports sprintf formatting}

\item{...}{Additional arguments for sprintf-style message formatting}
}
\description{
Logs a message at the INFO level, used for general informational messages
that highlight the progress of the application.
}
\examples{
logger.info("Application started successfully")
logger.info("Processing completed: \%d items processed", 58)

}
\seealso{
\code{\link{logger.log_level}}, \code{\link{INFO}}
}
