% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{logger.threshold}
\alias{logger.threshold}
\title{Current logging threshold}
\format{
An object of class \code{integer} of length 1.
}
\usage{
logger.threshold
}
\description{
Global variable that stores the current logging threshold level.
This variable is set by logger.init() and used by logger.log_level() to determine
whether messages should be displayed.
}
\keyword{datasets}
