% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.trace}
\alias{logger.trace}
\title{Log a trace level message}
\usage{
logger.trace(msg, ...)
}
\arguments{
\item{msg}{A character string containing the log message, supports sprintf formatting}

\item{...}{Additional arguments for sprintf-style message formatting}
}
\description{
Logs a message at the TRACE level, which is the most verbose logging level.
Trace messages are typically used for very detailed diagnostic information.
}
\examples{
logger.trace("Entering function with parameter: \%s", "value")
logger.trace("Variable state: x = \%d, y = \%s", 42, "test")

}
\seealso{
\code{\link{logger.log_level}}, \code{\link{TRACE}}
}
