% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pilot_client.R
\name{notifyDone}
\alias{notifyDone}
\title{Notify Execution Done}
\usage{
notifyDone(executionType)
}
\arguments{
\item{executionType}{Character string. The type of execution that has completed.
This parameter identifies what kind of operation or process has finished.}
}
\value{
No return value, called for side effects (logging notification)
}
\description{
Sends a notification indicating that a specific type of execution has completed.
This function is part of the HTTP fake client implementation for testing and
debugging purposes.
}
\details{
This is a fake client function used for testing and development. In the actual
implementation, this would typically send an HTTP notification to a server
endpoint. The fake implementation simply logs the notification using the
debug logging level.
}
\examples{
\dontrun{
notifyDone("data_processing")
notifyDone("model_training")
}

}
\seealso{
\code{\link{logger.debug}} for the logging function used
}
\keyword{internal}
