% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handler.R
\name{readInput}
\alias{readInput}
\title{Read Input Data}
\usage{
readInput(sourceFile)
}
\arguments{
\item{sourceFile}{Character string specifying the path to the input file}
}
\value{
The data object read from the RDS file
}
\description{
Reads input data from a source file. Currently supports RDS format only.
}
\examples{
\dontrun{
data <- readInput("input_data.rds")
}
}
