% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pilot_client.R
\name{storeConfiguration}
\alias{storeConfiguration}
\title{Store Configuration in Pilot}
\usage{
storeConfiguration(configuration)
}
\arguments{
\item{configuration}{The configuration object to be stored (will be converted to JSON)}
}
\value{
No return value. Called for its side effect of storing configuration.
}
\description{
Stores application configuration in the pilot service and updates the local
CONFIGURATION environment variable with the response.
}
\details{
If the HTTP_CLIENT_FAKE environment variable is set to "yes", the function
returns early without sending any request (useful for testing).
On successful storage, the function updates the CONFIGURATION environment
variable with the configuration returned from the server.
}
\examples{
\dontrun{
config <- list(setting1 = "value1", setting2 = 42)
storeConfiguration(config)
}
}
