% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handler.R
\name{storeResult}
\alias{storeResult}
\title{Store Result Data}
\usage{
storeResult(result, outputFile)
}
\arguments{
\item{result}{The R object to be stored}

\item{outputFile}{Character string specifying the path to the output file}
}
\value{
The return value from the underlying RDS storage function
}
\description{
Stores a result object to an output file in RDS format.
}
\examples{
\dontrun{
result <- list(data = c(1, 2, 3))
storeResult(result, "output.rds")
}
}
