% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{TRACE}
\alias{TRACE}
\title{Trace log level constant}
\format{
An object of class \code{integer} of length 1.
}
\usage{
TRACE
}
\description{
Numeric constant representing the TRACE log level with value 9.
Used for logging very detailed information, typically only of interest when diagnosing problems.
}
\keyword{datasets}
