% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{WARN}
\alias{WARN}
\title{Warning log level constant}
\format{
An object of class \code{integer} of length 1.
}
\usage{
WARN
}
\description{
Numeric constant representing the WARN log level with value 4.
Used for logging warning conditions that should be addressed but don't prevent execution.
}
\keyword{datasets}
